cache device uuid in client struct
When accessing/storing key info with userprefs, a device uuid is
required that makes it possible to distinguish between different
devices. On execution of lockdownd_client_new, the uuid is queried
via lockdown and now stored in the client struct for later reuse.
This patch also removes the uuid parameter from lockdownd_pair().