idevicebackup: correctly handle removal of existing .mddata + .mdinfo files
This commit fixes the problem that the .mddata and .mdinfo files were
not removed due to an incorrect check for the result of stat(). Since
the file data is sent in chunks we have to check for existing files
only when the first chunk of a file is received. That's why we now reset
hunk_index each time a file has been completely received.