Fix swig detection
AC_PROG_SWIG indicates failure to detect swig by setting the SWIG
variable to false. Commit f489971b6b moved the "$SWIG != false"
test before the call to AC_PROG_SWIG, which leads to the rest of
configure.ac to think swig has been successfully detected even
when not present. Move this test after the call to AC_PROG_SWIG.
This also gets rid of the SWIG_SUB variable which is not really
useful since the SWIG variable can be used for the same purpose.