idevicebackup2: Fix nasty "too long filename received" bug
If the device is sending files to the host, it sometimes requires a bit more
time to process them before sending. This appeared to happen mostly for larger
sqlite databases which appear to get some preprocessing on the device.
In such a "wait" situation, we receive no data and need to retry reading the
filename length again. Due to a code bug though which didn't reset the last
read length to zero, this length was incorrectly alternating between 1 and
16777216 due to the byte swapping. This ulitmativly lead to a broken backup
process. Now we properly wait for the device to preprocess any file before
sending the filename to the host.