installation_proxy: Fix operations exiting before being finished due to timeout
The internal status callback used a timeout of 30 seconds to receive status
messages about the progress of an operation. However, slow devices or large
app archives trigger this timeout causing the handler to return before
the operation was actually complete. This fixes it by removing the internal
timeout by waiting forever and only returning early in case of a real error.