idevicebackup2: Fix retry loop when device requests 0 files
Under certain circumstances a device could request an empty array
of filenames to be transferred with DLMessageDownloadFiles.
This was not handled correctly in idevicebackup2 by just doing
nothing, resulting in a 'retry loop' waiting endlessly for new
messages from the device.
The solution is to send a termination packet and a success
status response.

Thanks to pondahai for reporting.