lockdown: Fix error in re-pairing when previous record became invalid
When connecting with an existing pair record failed we attempted to
pair again which generated a new host id but then connect after
successful pairing tried to use host id from the old record and
failed. Make sure we forget the old host id when re-pairing.
This also fixes a possible memory leak.