Fix idevicedebug argv encoding
The argv encoding for the "A" command in idevicedebug could have an issue when encoding a character of 0x80 or higher (e.g. some UTF-8 characters) which could lead ot a crash in the on-device debugserver.  A simple fix to mask off the the high bits in debugserver_int2hex() addresses the issue.