idevicedebugserverproxy: Allow binding to any available port
To eliminate crosstalk between multiple proxies and their clients,
add support for binding to any free port provided by the OS to
idevicedebugserverproxy. To bind to any port, leave out the port
argument to idevicedebugserverproxy. In that case, the proxy will
print out a line with the port so clients can connect to it.

This is useful for a CI macOS host with multiple iDevices
connected, and where many independent tests each want their own
proxy instance connected to a particular device.