idevice: Restore original behavior for idevice_event_unsubscribe()
Due to an implementation detail from the past, a call to idevice_event_unsubscribe
would not cause the callback function to be called with IDEVICE_DEVICE_REMOVE
events, even though originally it was planned to be that way.
Due to the internal changes for the newer idevice_events_subscribe/unsubscribe
API, that behavior changed and it would call the callback.
Now to not break current software implementations depending on the original
behavior we modify the old API to behave the same as before the change.