common/userpref: remove X509V3_EXT_cleanup()
When building libimobiledevice with libressl 3.9 it fails with an
implicit declaration for X509V3_EXT_cleanup().

This was removed in libressl "because it was part of a dangerous,
incomplete and mostly useless API".

Additionally it doesn't seem to be actually used by libimobiledevice:

"It cleans up the stack containing custom extensions which
libimobiledevice doesn't use, it's only needed after calls to
X509V3_EXT_add{,_list,_alias}()."

Reference: https://github.com/libressl/portable/issues/1050
Signed-off-by: orbea <orbea@riseup.net>